Move indexes to separate module.
