+++ components/merkledb/src/indexes/proof_entry.rs
