+++ components/merkledb/src/indexes/proof_map_index/fig-index.png
