+++ components/merkledb/src/proto/mod.rs
@@ -21,7 +21,7 @@ use failure::Error;
 use protobuf::{well_known_types::Empty, RepeatedField};
 use std::borrow::Cow;
 
-use crate::{proof_map_index::ProofPath, BinaryKey, BinaryValue};
+use crate::{indexes::proof_map_index::ProofPath, BinaryKey, BinaryValue};
 use exonum_crypto::proto::*;
 use exonum_proto::ProtobufConvert;
 
@@ -125,8 +125,8 @@ mod tests {
     use std::fmt;
 
     use crate::{
-        access::AccessExt, proof_map_index::ToProofPath, proto, BinaryKey, BinaryValue, Database,
-        ListProof, MapProof, ObjectHash, TemporaryDB,
+        access::AccessExt, indexes::proof_map_index::ToProofPath, proto, BinaryKey, BinaryValue,
+        Database, ListProof, MapProof, ObjectHash, TemporaryDB,
     };
     use protobuf::RepeatedField;
